Moved OCI8Rows into rows.go
